PlayerFSM 관련 컨플릭트 해결
(cherry picked from commit 427538936012dca1118a2326456d64e38e79132e)

+++ Achromatic/Assets/Scripts/System/Manager/InputManager.cs
@@ -16,7 +16,7 @@ public class InputManager : SingletonBehavior<InputManager>
     const KeyCode FILTER = KeyCode.F;
 
     public UnityEvent JumpEvent;
-    public UnityEvent<float> MoveEvent;
+    public UnityEvent<int> MoveEvent;
     public UnityEvent<int> LookEvent;
     public UnityEvent<Vector2> DashEvent;
     public UnityEvent<Vector2> LightAttackEvent;
